% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.snreg.R
\name{coef.snreg}
\alias{coef.snreg}
\title{Extract Model Coefficients}
\usage{
\method{coef}{snreg}(obj, ...)
}
\arguments{
\item{obj}{an object of class \code{"snreg"}, typically returned by \code{\link{snreg}}.}

\item{...}{additional arguments (currently unused).}
}
\value{
A numeric vector containing the model coefficients.
}
\description{
\code{coef.snreg} is the S3 method for extracting the estimated regression
coefficients from an object of class \code{"snreg"}.
}
\details{
Coefficients from an snreg Model


This method simply returns the \code{coef} component stored inside the fitted
\code{"snreg"} object. If the object does not contain coefficient estimates
(e.g., if estimation was not completed in a scaffold), an informative error
is raised.
}
\examples{
\dontrun{
  m <- snreg(y ~ x1 + x2, data = df)
  coef(m)
}

}
